+++ src/components/MapStyleInput.svelte
@@ -198,8 +198,9 @@
           <option
             value={value.dropdownId}
             selected={dropdownValue.dropdownId === value.dropdownId}
-            >{value.text}</option
           >
+            {value.text}
+          </option>
         {/each}
       </optgroup>
     {/each}
@@ -230,7 +231,7 @@
   {/if}
 
   <div class="renderer-control">
-    <span>Rendered with</span>
+    <span class="nowrap">Rendered with</span>
     <select
       on:change={e => dispatch('selectRenderer', { value: e.target.value })}
     >
@@ -270,14 +271,28 @@
     color: white;
   }
 
+  .nowrap {
+    white-space: nowrap;
+  }
+
   .renderer-control {
     align-items: center;
     display: flex;
     flex-direction: row;
     gap: 0.5rem;
+    flex-wrap: nowrap;
   }
 
   .renderer-control select {
     flex-grow: 1;
+    min-width: 0px;
+  }
+
+  select,
+  option.selected {
+    display: block;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    overflow: hidden;
   }
 </style>