+++ src/components/MapsSwipeLayout.svelte
@@ -20,6 +20,12 @@
     if (!dragging || e.clientX === 0) return;
     sliderPosition = e.clientX - sliderWidth / 2;
   };
+
+  $: rightWidth = width - sliderPosition;
+  $: leftWidth = width - rightWidth;
+
+  $: themeLeftMapLabel = `right: unset; margin-right:unset; left:0; margin-left:1em; max-width:calc(${leftWidth}px - 6em)`;
+  $: themeRightMapLabel = `max-width:calc(${rightWidth}px - 6em)`;
 </script>
 
 <div
@@ -43,6 +49,7 @@
         numberOfMaps={maps.length}
         on:mapMove
         highlightDifferences={map.index === 1 && mapState?.showDiff}
+        labelStyle={map.index === 0 ? themeLeftMapLabel : themeRightMapLabel}
       />
     </div>
   {/each}