+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 0.13.1
+
 - Update dropdown styling to accomodate long style names
 - Add layer id to feature popup
 - Fix bug for branch dropdown option while polling a style